refactor: auth 도메인 레이어 패키지 분리
- RefreshToken 엔티티와 리포지토리를 domain/auth/로 이동하여
- game 모듈과 동일한 패키지 구조 적용
